requirements.txt

# Основные библиотеки
pandas>=2.0.0
numpy>=1.25.0
scikit-learn>=1.2.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
statsmodels>=0.14.0

# Работа с БД и SQL
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.9

# Работа с API
requests>=2.31.0

# Отчеты
reportlab>=4.0.0
pypdf
openpyxl>=3.1.2
xlsxwriter>=3.1.0

# Прочее
pyyaml>=6.0.0
joblib>=1.3.0
pytest>=7.4.0
pytest-cov>=4.1.0
loguru>=0.7.0

black

aiosmtpd>=1.4.4.post2
Pillow>=10.0.0
pyarrow>=15.0.0
kaleido>=0.2.1
python-dotenv>=1.0.1
